Add functionality to track if book has been read
User is now able to press button on book card to toggle whether a book
is read or unread. Each card will also update the corresponding book
object in myLibrary when toggled.

Additional logic added to form processing in order to account
for inputs with checkboxes, so that their values are added as True
if checkbox is checked, false if unchecked.

Form has additional checkbox field to check whether or not book is
read or unread when adding book to the library.